Reject api calls without valid access token
All calls to the api require an bearer token in the authorization header
I haven't checked all the edge cases yet, but this seems to do a fair
job of rejecting expired, malformed, and absent tokens. The angular app
uses an adal library which automatically intercepts requests and adds
this header to calls to this api.