Enable Controller/Method group-based authorization
Use @PreAuthorize annotation on controllers and/or methods to require
user to have certain group affiliation to access the resource. As
committed, reports controller requires user to be part of the 'admin'
group.

Note: groups must currently be specified by object id as this is what is
passed in via the id_token from azure